TiltModel/requirements.txt
Pillow==9.5.0
numpy==1.23.5
keras==2.12.0
tensorflow==2.12.0
opencv-python
pyautogui